update User Guide - Proposal and Grant Section
The previous user guide mentioned that this content type was a multi step process. This is incorrect as Table A has been updated and simplified. The table only asks for "individual funded effort (in months)" for "EPSCoR RII", in other words it only asks for your effort in the EPSCoR project, and not external grants. Similarly, R.gov wants to know your effort to the EPSCoR project. This information can be attained through the user profile.